scraper/requirements.txt
click
selenium
requests
pyyaml
tqdm
